voxblox protobuf update

+++ cblox/include/cblox/io/tsdf_submap_io.h
@@ -21,12 +21,12 @@ template<typename SubmapType>
 bool LoadSubmapFromStream(
     std::fstream* proto_file_ptr,
     typename SubmapCollection<SubmapType>::Ptr submap_collection_ptr,
-    uint32_t* tmp_byte_offset_ptr);
+    uint64_t* tmp_byte_offset_ptr);
 template<>
 bool LoadSubmapFromStream<cblox::TsdfSubmap>(
     std::fstream* proto_file_ptr,
     typename SubmapCollection<cblox::TsdfSubmap>::Ptr submap_collection_ptr,
-    uint32_t* tmp_byte_offset_ptr);
+    uint64_t* tmp_byte_offset_ptr);
 
 template <typename SubmapType>
 bool SaveSubmapCollection(
