+++ cblox/src/io/transformation_io.cpp
@@ -22,7 +22,7 @@ bool SaveTransformationArray(
   }
 
   // Write the total number of messages to the beginning of this file.
-  const uint32_t num_messages = transformation_array.size();
+  const uint64_t num_messages = transformation_array.size();
   if (!voxblox::utils::writeProtoMsgCountToStream(num_messages, &proto_file)) {
     LOG(ERROR) << "Could not write message number to file.";
     proto_file.close();
@@ -65,7 +65,7 @@ bool LoadTransformationArray(
   }
 
   // Unused byte offset result.
-  uint32_t tmp_byte_offset = 0u;
+  uint64_t tmp_byte_offset = 0u;
 
   // Get number of messages
   uint32_t num_protos;
